25/04/2023 - Carlos - Refactorizacion de codigo
