+++ webapp/src/components/map/OptionsPanel.tsx
@@ -7,7 +7,6 @@ import i18n from "../../i18n";
 import Mark from "./options/Mark";
 import {useSession} from "@inrupt/solid-ui-react";
 import MapSelector from "./options/MapSelector";
-import SelectorMap from "./SelectorMap";
 
 i18n.use(initReactI18next)
 
@@ -25,7 +24,7 @@ function OptionsPanel(props: { setItem: Function }) {
             <div id="optionsMenu">
                 <h1>{t("optionsMenu")}</h1>
                 <Search title={t("search")}/>
-                <SelectorMap setItem={props.setItem}/>
+                <MapSelector setItem={props.setItem}/>
                 <div id="filterDiv">
                     <Filter titleFilter={t("category")} nameFilter={"option"} usersWebId={user}
                             setItem={props.setItem}/>
